requirements.txt
pandas
openpyxl
python-dotenv
google-genai
openai
pypdf
python-docx
